:hammer: Refactor : usePhoto hook 분리
